Move settings into own package
