+++ src/main/kotlin/com/schwarzit/spectralIntellijPlugin/settings/ApplicationSettingsState.kt
@@ -1,4 +1,4 @@
-package com.schwarzit.spectralIntellijPlugin
+package com.schwarzit.spectralIntellijPlugin.settings
 
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.components.PersistentStateComponent
@@ -7,7 +7,10 @@ import com.intellij.openapi.components.Storage
 import com.intellij.util.xmlb.XmlSerializerUtil
 
 
-@State(name = "com.schwarzit.spectralIntellijPlugin.ApplicationSettingsState", storages = [Storage("spectral.xml")])
+@State(
+    name = "com.schwarzit.spectralIntellijPlugin.settings.ApplicationSettingsState",
+    storages = [Storage("spectral.xml")]
+)
 class ApplicationSettingsState : PersistentStateComponent<ApplicationSettingsState> {
 
     var ruleset: String = "https://raw.githubusercontent.com/SchwarzIT/api-linter-rules/main/spectral-api.yml"
